server/requirements.txt
aioredis==2.0.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==1.16.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.65.1
fastapi-restful==0.3.2
future==0.18.2
h11==0.12.0
httpcore==0.13.3
httpx==0.18.1
idna==3.2
importlib-metadata==4.5.0
iniconfig==1.1.1
motor==2.4.0
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==3.5.3
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
treelib==1.6.1
typing-extensions==3.10.0.0
uuid==1.30
uvicorn==0.14.0
zipp==3.4.1
